Added Synty assets and created project folder heirarchy
Some materials in the added asset packs may still be broken but most have been updated to be URP materials. To view the different asset offerings in each pack, simply navigate to their folder in the 'Synty Assets' folder and find a sample scene for each pack.

Moving forward, please try not to alter the files in the Synty Assets folder to prevent merge conflicts.